riscv64: auto generate

+++ libc_musl.go
@@ -647,7 +647,7 @@ func ___randname(tls *TLS, template uintptr) (r1 uintptr) {
 	X__clock_gettime(tls, CLOCK_REALTIME, bp)
 	goto _2
 _2:
-	r = uint64((*(*Ttimespec)(unsafe.Pointer(bp))).Ftv_sec+int64((*(*Ttimespec)(unsafe.Pointer(bp))).Ftv_nsec)) + uint64(tls.ID)*uint64(65537)
+	r = uint64((*(*Ttimespec)(unsafe.Pointer(bp))).Ftv_sec+(*(*Ttimespec)(unsafe.Pointer(bp))).Ftv_nsec) + uint64(tls.ID)*uint64(65537)
 	i = 0
 	for {
 		if !(i < int32(6)) {
